backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai==1.52.1
python-dotenv==1.0.1
